+++ classes/operation/get_course_completion_criteria.php
@@ -1,0 +1,32 @@
+<?php
+// This file is part of Moodle - http://moodle.org/
+//
+// Moodle is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+/**
+ * Course completion criteria read operation.
+ *
+ * @package    local_moodlia
+ * @copyright  2026 Pablo Gallego
+ * @license    https://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+
+namespace local_moodlia\operation;
+
+/**
+ * Returns the global completion criteria configured for a course.
+ */
+class get_course_completion_criteria {
+    /**
+     * Execute the operation.
+     *
+     * @param int $courseid Courseid.
+     * @return array
+     */
+    public static function execute(int $courseid): array {
+        return course_completion_tools::configuration_to_response(course_tools::get_course($courseid));
+    }
+}